chore(codegen): move CrossRegionCopying to separate plugin
