+++ codegen/smithy-aws-typescript-codegen/src/main/java/software/amazon/smithy/aws/typescript/codegen/AddBuiltinPlugins.java
@@ -39,15 +39,7 @@ import software.amazon.smithy.utils.SetUtils;
  * Adds all built-in runtime client plugins to clients.
  */
 public class AddBuiltinPlugins implements TypeScriptIntegration {
-
     private static final Set<String> ROUTE_53_ID_MEMBERS = SetUtils.of("DelegationSetId", "HostedZoneId", "Id");
-
-    private static final Set<String> RDS_PRESIGNED_URL_OPERATIONS = SetUtils.of(
-        "CopyDBSnapshot",
-        "CreateDBInstanceReadReplica",
-        "CreateDBCluster",
-        "CopyDBClusterSnapshot"
-    );
 
     @Override
     public List<RuntimeClientPlugin> getClientPlugins() {
@@ -103,12 +95,6 @@ public class AddBuiltinPlugins implements TypeScriptIntegration {
                                          "ChangeResourceRecordSets", HAS_MIDDLEWARE)
                         .operationPredicate((m, s, o) -> o.getId().getName().equals("ChangeResourceRecordSets")
                                             && testServiceId(s, "Route 53"))
-                        .build(),
-                RuntimeClientPlugin.builder()
-                        .withConventions(AwsDependency.RDS_MIDDLEWARE.dependency, "CrossRegionPresignedUrl",
-                                         HAS_MIDDLEWARE)
-                        .operationPredicate((m, s, o) -> RDS_PRESIGNED_URL_OPERATIONS.contains(o.getId().getName())
-                                            && testServiceId(s, "RDS"))
                         .build(),
                 RuntimeClientPlugin.builder()
                         .withConventions(AwsDependency.ROUTE53_MIDDLEWARE.dependency, "IdNormalizer",
