+++ codegen/smithy-aws-typescript-codegen/src/main/resources/META-INF/services/software.amazon.smithy.typescript.codegen.integration.TypeScriptIntegration
@@ -15,3 +15,4 @@ software.amazon.smithy.aws.typescript.codegen.AddTranscribeStreamingDependency
 software.amazon.smithy.aws.typescript.codegen.AddUserAgentDependency
 software.amazon.smithy.aws.typescript.codegen.AddOmitRetryHeadersDependency
 software.amazon.smithy.aws.typescript.codegen.StripNewEnumNames
+software.amazon.smithy.aws.typescript.codegen.AddCrossRegionCopyingPlugin
